--- server/bin/pbench-report-status
@@ -1,1 +1,0 @@
-pbench-trampoline
