--- server/bin/state/test-23.config/pbench-server.cfg
@@ -1,51 +1,0 @@
-install-dir = %(unittest-dir)s/opt/pbench-server
-
-###########################################################################
-## Deployment section
-###########################################################################
-[pbench-server]
-pbench-top-dir = %(unittest-dir)s/pbench
-pbench-local-dir = %(unittest-dir)s/pbench-local
-pbench-backup-dir = %(pbench-local-dir)s/archive.backup
-# Add role for sync'ing with satellites
-roles = pbench-maintenance, pbench-results, pbench-sync-satellites
-debug_unittest = True
-commit_id = unit-test
-environment = unit-test
-admin-email = unit-test-user@example.com
-
-[Indexing]
-server = elasticsearch.example.com:9280
-# No index_prefix so logger should be used in pbench-report-status
-
-[apache]
-documentroot = %(unittest-dir)s/var-www-html
-
-[logging]
-logger_type = file
-
-###########################################################################
-# crontab roles
-
-# The definition of the crontab role for sync'ing satellite pbench servers.
-[pbench-sync-satellites]
-host = %(default-host)s
-satellites = satellite-one
-tasks = pbench-sync
-
-# Template values for this satellite
-[satellite-one]
-# NOTE WELL: this satellite host name, pbench-satellite.example.com, is the
-# host name expected by the mock ssh command which triggers the behavior of
-# running the actual commands instead of just echoing them.
-satellite-host = pbench-satellite.example.com
-satellite-prefix = ONE
-satellite-lock = pbench-sync-satellite-%(satellite-prefix)s.lock
-satellite-archive = %(unittest-dir)s/pbench-satellite/archive/fs-version-001
-satellite-opt = %(unittest-dir)s/opt/pbench-server-satellite
-
-###########################################################################
-# The rest will come from the default config file.
-[config]
-path = %(install-dir)s/lib/config
-files = pbench-server-default.cfg
