--- server/bin/state/test-23.setup
@@ -1,1 +1,0 @@
-test-12.setup
