--- server/bin/test-bin/curl
@@ -1,18 +1,0 @@
-#! /bin/bash
-
-echo "${0##*/} ${*}" | sed 's/pbench-report-status.[0-9][0-9]*/pbench-report-status.NNNN/' >> ${_testlog}
-
-file=""
-while [ ! -z "${1}" ]; do
-    if [ "${1}" != "--data" ]; then
-        shift
-        continue 
-    fi
-    shift
-    file=${1#@}
-done
-if [ ! -z "${file}" -a -e "${file}" ]; then
-    cat ${file} >> ${_testcurlpayload}
-fi
-printf "201"
-exit 0
