+++ server/bin/unittests
@@ -213,11 +213,6 @@ function _dump_logs {
         echo "--- $(basename $_testlog) file contents" >> $_testout
     fi
 
-    if [ -s $_testcurlpayload ]; then
-        echo "+++ $(basename $_testcurlpayload) file contents" >> $_testout
-        _normalize_output $_testcurlpayload >> $_testout 2>&1
-        echo "--- $(basename $_testcurlpayload) file contents" >> $_testout
-    fi
 
     if [ -s $_testloggerpayload ]; then
         echo "+++ $(basename $_testloggerpayload) file contents" >> $_testout
@@ -486,7 +481,6 @@ function _run_test {
     export _testactout=$_testroot/actoutput.txt
     export _testactlog=$_testroot/test-activation-execution.log
     export _testlog=$_testroot/test-execution.log
-    export _testcurlpayload=$_testroot/test-curl-payload.log
     export _testloggerpayload=$_testroot/test-logger-payload.log
     export _testdir=$_testroot/pbench
     export _testdir_local=$_testroot/pbench-local
@@ -533,7 +527,6 @@ function _reset_state {
     rm -f $_testactout
     rm -f $_testlog
     rm -f $_testactlog
-    rm -f $_testcurlpayload
     rm -f $_testloggerpayload
     rm -rf $_testdir
     if [[ -d $_testdir ]]; then
@@ -704,9 +697,6 @@ declare -A cmds=(
 
     # Actually exercises satellite code
     [test-22]="_run pbench-satellite-cleanup"
-
-    # Uses pbench-dispatch to test report status via syslog
-    [test-23]="_run pbench-dispatch"
 
     # Tests that the behavior of the find command works as we expect to
     # ensure that buckets for pbench-unpack-tarballs work properly.
